add js/tsdoc and adjust some of the docs
