+++ docs/content/contents.mdx
@@ -3,6 +3,8 @@ course:
   title: "Might Fail"
   parts:
     - title: "Getting Started"
+      slug: getting-started
+      filePath: /getting-started.mdx
       icon: "rocket"
       pages:
         - title: "❌ try, catch, finally"
@@ -29,4 +31,30 @@ course:
           filePath: /might-and-fail.mdx
           slug: might-and-fail
           type: documentation
+    - title: "API Reference"
+      slug: api-reference
+      filePath: /jsdocs/variables/default.mdx
+      icon: "file-text"
+      pages:
+        - title: "`mightFail`"
+          filePath: /jsdocs/functions/mightFail.mdx
+          slug: api-might-fail
+          type: documentation
+        - title: "`mightFailSync`"
+          filePath: /jsdocs/functions/mightFailSync.mdx
+          slug: api-might-fail-sync
+          type: documentation
+        - title: "`Either`"
+          filePath: /jsdocs/type-aliases/Either.mdx
+          slug: api-either
+          type: documentation
+        - title: "`Fail`"
+          filePath: /jsdocs/functions/Fail.mdx
+          slug: api-fail
+          type: documentation
+        - title: "`Might`"
+          filePath: /jsdocs/functions/Might.mdx
+          slug: api-might
+          type: documentation
+          
 ---
