+++ docs/content/jsdocs/functions/mightFailFunction.mdx
@@ -1,0 +1,21 @@
+# Function: mightFailFunction()
+
+```ts
+function mightFailFunction<T>(promise: Promise<T>): Promise<Either<T>>
+```
+
+## Type Parameters
+
+| Type Parameter |
+| ------ |
+| `T` |
+
+## Parameters
+
+| Parameter | Type |
+| ------ | ------ |
+| `promise` | `Promise`\<`T`\> |
+
+## Returns
+
+`Promise`\<[`Either`](https://mightfail.dev/type-aliases/Either.mdx)\<`T`\>\>
