+++ docs/content/jsdocs/index.mdx
@@ -1,0 +1,29 @@
+# might-fail
+
+## References
+
+### makeMightFail
+
+Re-exports makeMightFail
+
+***
+
+### makeMightFailSync
+
+Re-exports makeMightFailSync
+
+## Type Aliases
+
+- [Either](https://mightfail.dev/type-aliases/Either.mdx)
+
+## Variables
+
+- [default](https://mightfail.dev/variables/default.mdx)
+
+## Functions
+
+- [Fail](https://mightfail.dev/functions/Fail.mdx)
+- [Might](https://mightfail.dev/functions/Might.mdx)
+- [mightFail](https://mightfail.dev/functions/mightFail.mdx)
+- [mightFailFunction](https://mightfail.dev/functions/mightFailFunction.mdx)
+- [mightFailSync](https://mightfail.dev/functions/mightFailSync.mdx)
