+++ docs/content/mightFail.mdx
@@ -2,35 +2,39 @@
 
 # `mightFail` Async
 
-## Wrap Promise in `mightFail`
+## Wrap Promise
+
+Simply wrap any promise in `mightFail` and await the result. 
+
+You will get and `Either` object that contains either an error or the result. Never both. 
+
+```ts
+const EitherObject = await mightFail(fetch("/posts"));
+```
+
+But **don't** store the `Either` object directly. Instead, destructure it in the same statement.
+
+```ts
+// tuple destructuring
+const [ error, result ] = await mightFail(fetch("/posts"));
+
+// object destructuring
+const { error, result } = await mightFail(fetch("/posts"));
+
+// go destructuring
+// import { mightFail } from "might-fail/go";
+const [ result, error ] = await mightFail(fetch("/posts"));
+```
+
+You can choose to destructure this object as a tuple or as an object. Or as a backwards tuple if you prefer it that way.
+
+We think that the **tuple** option is the best, but you do you.
+
+## Guard
+
+Once you've awaited and destructured the `Either` object, use guard clauses to handle the error, and handle the success case at the end.
 
 <Tabs>
-<Tab name="object">
-```ts
-const { error: networkError, result } = await mightFail(fetch("/posts"));
-
-if (networkError) {
-  // handle network error
-  return;
-}
-
-if (!result.ok) {
-  // handle an error response from server
-  return;
-}
-
-const { error: convertToJSONError, result: posts } = await mightFail(
-  result.json()
-);
-
-if (convertToJSONError) {
-  // handle convertToJSONError
-  return;
-}
-
-posts.map((post) => console.log(post.title));
-```
-</Tab>
 
 <Tab name="tuple">
 ```ts
@@ -58,6 +62,34 @@ if (convertToJSONError) {
 posts.map((post) => console.log(post.title));
 ```
 </Tab>
+
+<Tab name="object">
+```ts
+const { error: networkError, result } = await mightFail(fetch("/posts"));
+
+if (networkError) {
+  // handle network error
+  return;
+}
+
+if (!result.ok) {
+  // handle an error response from server
+  return;
+}
+
+const { error: convertToJSONError, result: posts } = await mightFail(
+  result.json()
+);
+
+if (convertToJSONError) {
+  // handle convertToJSONError
+  return;
+}
+
+posts.map((post) => console.log(post.title));
+```
+</Tab>
+
 <Tab name="go">
 ```ts
 const [ result, networkError ] = await mightFail(fetch("/posts"));
@@ -86,13 +118,15 @@ posts.map((post) => console.log(post.title));
 </Tab>
 </Tabs>
 
-## Or Wrap Async Function in `makeMightFail`
+## `makeMightFail`
+
+You also have the option to make your own `mightFail` wrapper function. Pass in any async function to `makeMightFail` and it will return a `mightFail` wrapper function for you to use.
 
 <Tabs>
-<Tab name="object">
+<Tab name="tuple">
 ```ts
 const get = makeMightFail(axios.get);
-const { error, result } = await get("/posts");
+const [ error, result ] = await get("/posts");
 
 if (error) {
   // handle error
@@ -103,10 +137,10 @@ const posts = result.data
 posts.map((post) => console.log(post.title));
 ```
 </Tab>
-<Tab name="tuple">
+<Tab name="object">
 ```ts
 const get = makeMightFail(axios.get);
-const [ error, result ] = await get("/posts");
+const { error, result } = await get("/posts");
 
 if (error) {
   // handle error
