+++ package.json
@@ -21,7 +21,8 @@
     "build:esm": "tsc",
     "build:cjs": "tsc --project tsconfig.cjs.json",
     "test": "vitest",
-    "publish": "npm run build && npm publish"
+    "publish": "npm run build && npm publish",
+    "typedoc": "typedoc src"
   },
   "files": [
     "dist"
@@ -47,6 +48,9 @@
     "typescript"
   ],
   "devDependencies": {
+    "jsdoc-to-markdown": "^9.0.2",
+    "typedoc": "^0.26.8",
+    "typedoc-plugin-markdown": "^4.2.9",
     "typescript": "^5.6.2",
     "vitest": "^2.1.1"
   }
