+++ src/mightFail.ts
@@ -1,9 +1,11 @@
 import { type Either } from "./Either"
-import { createEither, handleError, makeProxyHandler } from "./utils"
-import { MightFail, MightFailFunction, NonUndefined } from "./utils.type"
+import { makeProxyHandler } from "./utils/staticMethodsProxy"
+import { handleError } from "./utils/errors"
+import { createEither } from "./utils/createEither"
+import { MightFail, MightFailFunction, NonUndefined } from "./utils/utils.type"
 
 export const mightFailFunction: MightFailFunction<"standard"> = async function <T>(
-  promise: Promise<T>,
+  promise: Promise<T>
 ): Promise<Either<T>> {
   try {
     const result = await promise
@@ -20,13 +22,12 @@ export const mightFailFunction: MightFailFunction<"standard"> = async function <
  * either contains a 'result' of type T if the promise resolves successfully, or an 'error' of type Error
  * if the promise is rejected.
  *
- * @export
  * @template T The type of the result value.
  * @param {Promise<T>} promise - The promise to be wrapped in an Either. This is an asynchronous operation that
  * should resolve with a value of type T or reject with an Error.
- * @return {Promise<Either<T>>} A Promise that resolves with an Either. This Either is a Success<T> with
+ * @return {Promise<Either<T>>} A Promise that resolves with an Either. This Either is a `Success<T>` with
  * the 'result' property set to the value resolved by the promise if successful, and 'error' as undefined.
- * In case of failure, it's a Failure with 'result' as undefined and 'error' of type Error. `error` will **always** be an instance of Error.
+ * In case of failure, it's a `Failure` with 'result' as undefined and 'error' of type Error. `error` will **always** be an instance of Error.
  *
  * @example
  * // Example of wrapping an async function that might fail:
@@ -48,7 +49,7 @@ export const mightFailFunction: MightFailFunction<"standard"> = async function <
  */
 export const mightFail: MightFail<"standard"> = new Proxy(
   mightFailFunction,
-  makeProxyHandler(mightFailFunction),
+  makeProxyHandler(mightFailFunction)
 ) as MightFail<"standard">
 
 /**
@@ -57,13 +58,12 @@ export const mightFail: MightFail<"standard"> = new Proxy(
  * It returns an object that either contains a 'result' of type T if the function succeeds,
  * or an 'error' of type Error if the function throws an error.
  *
- * @export
- * @template T The type of the result value.
+ * @template T The type of the result value.◊
  * @param {() => T} func - A wrapper function that is expected to invoke the throwing function.
  *  That function should return a value of type T or throw an error.
- * @return {Either<T>} An object that is either a Success<T> with the result property set to the value returned by `func`,
- *                     or a Failure with the error property set to the caught error. Success<T> has a 'result' of type T
- *                     and 'error' as null. Failure has 'result' as null and 'error' of type Error.
+ * @return {Either<T>} An object that is either a `Success<T>` with the result property set to the value returned by `func`,
+ *                     or a `Failure` with the error property set to the caught error. `Success<T>` has a 'result' of type T
+ *                     and 'error' as null. `Failure` has 'result' as null and 'error' of type Error.
  * @example
  * // Example of wrapping a synchronous function that might throw an error:
  * const {error, result} = mightFailSync(() => JSON.parse(""));
@@ -74,8 +74,7 @@ export const mightFail: MightFail<"standard"> = new Proxy(
  * }
  * console.log('Parsed object:', result);
  */
-
-export function mightFailSync<T>(func: () => T): Either<T> {
+export const mightFailSync = function mightFailSync<T>(func: () => T): Either<T> {
   try {
     const result = func()
     return createEither<T>({ error: undefined, result })
@@ -86,23 +85,22 @@ export function mightFailSync<T>(func: () => T): Either<T> {
 }
 
 /**
- * A pure constructor function that takes a non-null value and returns an Either object with the value as the result and undefined as the error.
+ * A pure constructor function that takes a non-null value and returns an `Either` object with the value as the result and undefined as the error.
  *
  * @param result
- * @constructor
  */
-export function Might<T>(result: NonUndefined<T>): Either<T> {
+export const Might = function Might<T>(result: NonUndefined<T>): Either<T> {
   return createEither<T>({ result, error: undefined })
 }
 
 /**
- * A constructor function that takes an error and returns an Either object with undefined as the result and the error as the error.
+ * A constructor function that takes an error and returns an `Either` object with undefined as the result and the error as the error.
  *
  * The error will **always** be an instance of Error.
  *
  * @param error
- * @constructor
  */
-export function Fail(error: unknown): Either<undefined> {
+export const Fail = function Fail(error: unknown): Either<undefined> {
   return createEither<undefined>({ result: undefined, error: handleError(error) })
 }
+
